split into files, misc updates
