+++ Community/Eos Lighting/automation/jsr223/python/community/290_eos.py
@@ -16,14 +16,7 @@ def eos_rule_reinit(event):
 def eos_rule_scene_update(event):
     """Eos Scene Received Update Rule"""
     eos.log.debug("{rule} triggered by '{name}' with scene '{scene}'".format(rule=eos.RULE_SCENE_NAME, name=event.itemName, scene=event.itemCommand))
-
-    item = itemRegistry.get(event.itemName)
-
-    for light in eos.get_light_items(eos.get_item_eos_group(item)):
-        try: eos.update_light(light)
-        except: continue
-    for group in eos.get_group_items(eos.get_item_eos_group(item)):
-        eos.update_group(group, True)
+    eos.update_scene(itemRegistry.get(event.itemName))
 
 
 def eos_rule_light_update(event):
@@ -35,7 +28,7 @@ def eos_rule_light_update(event):
 def eos_rule_level_source_update(event):
     """Eos Level Source Received Update Rule"""
     eos.log.debug("{rule} triggered by '{name}' with state '{state}'".format(rule=eos.RULE_LEVEL_SOURCE_NAME, name=event.itemName, state=event.itemState))
-    eos.update_group(itemRegistry.get(eos.master_group_name))
+    eos.update_group(itemRegistry.get(eos.config.master_group_name()))
 
 
 @log_traceback
